Add Font file and suppress a warning

+++ Geometry.cpp
@@ -40,10 +40,10 @@ int orientation(Point p, Point q, Point r)
 {
 	// See 10th slides from following link for derivation of the formula
 	// http://www.dcs.gla.ac.uk/~pat/52233/slides/Geometry1x1.pdf
-	int val = (q.y - p.y) * (r.x - q.x) -
+	double val = (q.y - p.y) * (r.x - q.x) -
 		(q.x - p.x) * (r.y - q.y);
 
-	if (val == 0) return 0;  // colinear
+	if (int(val) == 0) return 0;  // colinear
 
 	return (val > 0) ? 1 : 2; // clock or counterclock wise
 }
